refactor(query_api): extract memory CRUD methods to memory.rs
Move memory_create/update/mark_obsolete/search/for_symbol/for_path/for_edges/
evidence_for_symbol_and_edges/for_call_path_hash/rebind/validate/doctor/list/get
(all thin delegators to crate::query::memory) into memory.rs. memory_anchor_health
stays in mod.rs (it's index-status, not CRUD). Also restores important_symbols's
doc comment, which the prior importance extraction left behind in mod.rs (it had
silently re-attached to memory_create). No behavior change.
