+++ crates/rag-rat-core/src/index/query_api/memory.rs
@@ -1,0 +1,125 @@
+use super::*;
+
+impl IndexDatabase {
+    pub fn memory_create(
+        &self,
+        request: crate::query::memory::RepoMemoryCreate,
+    ) -> anyhow::Result<crate::query::memory::RepoMemoryCreateResult> {
+        crate::query::memory::create_memory(self.storage.connection(), request)
+    }
+
+    pub fn memory_update(
+        &self,
+        update: crate::query::memory::RepoMemoryUpdate,
+    ) -> anyhow::Result<crate::query::memory::RepoMemory> {
+        crate::query::memory::update_memory(self.storage.connection(), update)
+    }
+
+    pub fn memory_mark_obsolete(
+        &self,
+        memory_id: &str,
+    ) -> anyhow::Result<crate::query::memory::RepoMemory> {
+        crate::query::memory::mark_obsolete(self.storage.connection(), memory_id)
+    }
+
+    pub fn memory_search(
+        &self,
+        query: &str,
+        limit: u32,
+    ) -> anyhow::Result<Vec<crate::query::memory::RepoMemory>> {
+        crate::query::memory::memory_search(self.storage.connection(), query, limit)
+    }
+
+    pub fn memory_for_symbol(
+        &self,
+        symbol: &crate::query::symbol::SymbolHit,
+        limit: u32,
+    ) -> anyhow::Result<Vec<crate::query::memory::RepoMemory>> {
+        crate::query::memory::memories_for_symbol(self.storage.connection(), symbol, limit)
+    }
+
+    pub fn memory_for_path(
+        &self,
+        path: &str,
+        limit: u32,
+    ) -> anyhow::Result<Vec<crate::query::memory::RepoMemory>> {
+        crate::query::memory::memories_for_path(self.storage.connection(), path, limit)
+    }
+
+    pub fn memory_for_edges(
+        &self,
+        edge_ids: &[i64],
+        limit: u32,
+    ) -> anyhow::Result<Vec<crate::query::memory::RepoMemory>> {
+        crate::query::memory::memories_for_edges(self.storage.connection(), edge_ids, limit)
+    }
+
+    pub fn memory_evidence_for_symbol_and_edges(
+        &self,
+        symbol: &crate::query::symbol::SymbolHit,
+        caller_edge_ids: &[i64],
+        callee_edge_ids: &[i64],
+        limit: u32,
+    ) -> anyhow::Result<crate::query::memory::RepoMemoryEvidence> {
+        // This wrapper exposes only the evidence; the impact builder consumes the truncation flag
+        // directly from the core fn.
+        crate::query::memory::memory_evidence_for_symbol_and_edges(
+            self.storage.connection(),
+            symbol,
+            caller_edge_ids,
+            callee_edge_ids,
+            limit,
+        )
+        .map(|(evidence, _truncated)| evidence)
+    }
+
+    pub fn memory_for_call_path_hash(
+        &self,
+        edge_sequence_hash: &str,
+        limit: u32,
+    ) -> anyhow::Result<Vec<crate::query::memory::RepoMemory>> {
+        crate::query::memory::memories_for_call_path_hash(
+            self.storage.connection(),
+            edge_sequence_hash,
+            limit,
+        )
+    }
+
+    pub fn memory_rebind(
+        &self,
+        memory_id: &str,
+        bind: crate::query::memory::RepoMemoryBindTarget,
+    ) -> anyhow::Result<crate::query::memory::RepoMemory> {
+        crate::query::memory::rebind_memory(self.storage.connection(), memory_id, bind)
+    }
+
+    pub fn memory_validate(
+        &self,
+    ) -> anyhow::Result<crate::query::memory::RepoMemoryValidationReport> {
+        crate::query::memory::validate_memories(
+            self.storage.connection(),
+            self.storage.source_root(),
+        )
+    }
+
+    pub fn memory_doctor(&self) -> anyhow::Result<Vec<crate::query::memory::MemoryDoctorEntry>> {
+        crate::query::memory::doctor_report(self.storage.connection())
+    }
+
+    /// Read-only list of active+stale memories, optionally filtered by binding_kind.
+    /// `kind` filters by binding kind (e.g. `Some("dir")`); `None` returns all.
+    pub fn memory_list(
+        &self,
+        kind: Option<&str>,
+    ) -> anyhow::Result<Vec<crate::query::memory::MemorySummary>> {
+        crate::query::memory::list_memories(self.storage.connection(), kind)
+    }
+
+    /// Fetch a single memory by id, returning `None` when not found.
+    pub fn memory_get(
+        &self,
+        memory_id: &str,
+    ) -> anyhow::Result<Option<crate::query::memory::RepoMemory>> {
+        crate::query::memory::memory_by_id(self.storage.connection(), memory_id)
+    }
+}
