+++ crates/rag-rat-core/src/index/query_api/mod.rs
@@ -6,6 +6,7 @@ use crate::index::oracle::{
 };
 
 mod importance;
+mod memory;
 
 pub use importance::ImportantSymbolsRequest;
 
@@ -1857,151 +1858,6 @@ impl IndexDatabase {
         options: crate::query::clusters::RepoClustersOptions,
     ) -> anyhow::Result<crate::query::clusters::RepoClustersReport> {
         crate::query::clusters::repo_clusters(self.storage.connection(), options)
-    }
-
-    /// Top load-bearing symbols by weighted PageRank over the active checkout's edge graph (#108).
-    /// `personalize_to` biases importance toward those symbol ids (changed/query symbols); empty =
-    /// global.
-    ///
-    /// When a SCIP oracle run exists for this checkout, ranking uses the compiler-verified graph
-    /// (contradicted edges dropped, upgrades retargeted, confirmed/upgraded edges weighted above
-    /// heuristic) — otherwise the heuristic graph with confidence weighting. The oracle lookup is
-    /// gated on a run existing, so absent oracle data it costs nothing (no scan).
-    /// Rank load-bearing symbols, returning the labeled [`ImportantSymbolsResult`] (mode + seed
-    /// provenance), per the spec's "three scales". Seed resolution happens HERE, at the query
-    /// boundary, because it needs both the symbol index (name/path → id) and git (the working-set
-    /// diff) — `query::pagerank` stays a pure ranking primitive over raw ids.
-    ///
-    /// Seed precedence:
-    /// - explicit `request.personalize` (names / paths / numeric ids) → `SeedKind::Explicit`;
-    /// - else, if `request.auto_seed_from_diff` (the MCP default), the current git diff →
-    ///   `SeedKind::GitDiff`;
-    /// - else (the CLI default, or an explicit empty/`global` selector) → global, un-seeded.
-    ///
-    /// A seed intent that resolves to NO in-graph symbol (bad names only, or a diff with no indexed
-    /// symbols) does NOT hard-error: it falls through to global ranking but REPORTS the
-    /// fall-through (`mode = global …` + `reason` + the diff counts), so the caller sees WHY it
-    /// was un-seeded.
-    pub fn memory_create(
-        &self,
-        request: crate::query::memory::RepoMemoryCreate,
-    ) -> anyhow::Result<crate::query::memory::RepoMemoryCreateResult> {
-        crate::query::memory::create_memory(self.storage.connection(), request)
-    }
-
-    pub fn memory_update(
-        &self,
-        update: crate::query::memory::RepoMemoryUpdate,
-    ) -> anyhow::Result<crate::query::memory::RepoMemory> {
-        crate::query::memory::update_memory(self.storage.connection(), update)
-    }
-
-    pub fn memory_mark_obsolete(
-        &self,
-        memory_id: &str,
-    ) -> anyhow::Result<crate::query::memory::RepoMemory> {
-        crate::query::memory::mark_obsolete(self.storage.connection(), memory_id)
-    }
-
-    pub fn memory_search(
-        &self,
-        query: &str,
-        limit: u32,
-    ) -> anyhow::Result<Vec<crate::query::memory::RepoMemory>> {
-        crate::query::memory::memory_search(self.storage.connection(), query, limit)
-    }
-
-    pub fn memory_for_symbol(
-        &self,
-        symbol: &crate::query::symbol::SymbolHit,
-        limit: u32,
-    ) -> anyhow::Result<Vec<crate::query::memory::RepoMemory>> {
-        crate::query::memory::memories_for_symbol(self.storage.connection(), symbol, limit)
-    }
-
-    pub fn memory_for_path(
-        &self,
-        path: &str,
-        limit: u32,
-    ) -> anyhow::Result<Vec<crate::query::memory::RepoMemory>> {
-        crate::query::memory::memories_for_path(self.storage.connection(), path, limit)
-    }
-
-    pub fn memory_for_edges(
-        &self,
-        edge_ids: &[i64],
-        limit: u32,
-    ) -> anyhow::Result<Vec<crate::query::memory::RepoMemory>> {
-        crate::query::memory::memories_for_edges(self.storage.connection(), edge_ids, limit)
-    }
-
-    pub fn memory_evidence_for_symbol_and_edges(
-        &self,
-        symbol: &crate::query::symbol::SymbolHit,
-        caller_edge_ids: &[i64],
-        callee_edge_ids: &[i64],
-        limit: u32,
-    ) -> anyhow::Result<crate::query::memory::RepoMemoryEvidence> {
-        // This wrapper exposes only the evidence; the impact builder consumes the truncation flag
-        // directly from the core fn.
-        crate::query::memory::memory_evidence_for_symbol_and_edges(
-            self.storage.connection(),
-            symbol,
-            caller_edge_ids,
-            callee_edge_ids,
-            limit,
-        )
-        .map(|(evidence, _truncated)| evidence)
-    }
-
-    pub fn memory_for_call_path_hash(
-        &self,
-        edge_sequence_hash: &str,
-        limit: u32,
-    ) -> anyhow::Result<Vec<crate::query::memory::RepoMemory>> {
-        crate::query::memory::memories_for_call_path_hash(
-            self.storage.connection(),
-            edge_sequence_hash,
-            limit,
-        )
-    }
-
-    pub fn memory_rebind(
-        &self,
-        memory_id: &str,
-        bind: crate::query::memory::RepoMemoryBindTarget,
-    ) -> anyhow::Result<crate::query::memory::RepoMemory> {
-        crate::query::memory::rebind_memory(self.storage.connection(), memory_id, bind)
-    }
-
-    pub fn memory_validate(
-        &self,
-    ) -> anyhow::Result<crate::query::memory::RepoMemoryValidationReport> {
-        crate::query::memory::validate_memories(
-            self.storage.connection(),
-            self.storage.source_root(),
-        )
-    }
-
-    pub fn memory_doctor(&self) -> anyhow::Result<Vec<crate::query::memory::MemoryDoctorEntry>> {
-        crate::query::memory::doctor_report(self.storage.connection())
-    }
-
-    /// Read-only list of active+stale memories, optionally filtered by binding_kind.
-    /// `kind` filters by binding kind (e.g. `Some("dir")`); `None` returns all.
-    pub fn memory_list(
-        &self,
-        kind: Option<&str>,
-    ) -> anyhow::Result<Vec<crate::query::memory::MemorySummary>> {
-        crate::query::memory::list_memories(self.storage.connection(), kind)
-    }
-
-    /// Fetch a single memory by id, returning `None` when not found.
-    pub fn memory_get(
-        &self,
-        memory_id: &str,
-    ) -> anyhow::Result<Option<crate::query::memory::RepoMemory>> {
-        crate::query::memory::memory_by_id(self.storage.connection(), memory_id)
     }
 }
